scantailor EXPERIMENTAL-2016-02-22 (devel)
Update :devel and :head to point to the author's newest work.

Fix :stable by using fully qualified boost::lambda::bind. As the
upstream commit fixing this says, "With boost 1.60 there's a namespace
conflict between boost::bind and boost::lambda::bind and placeholders
are no longer in global namespace, so use fully qualified names for
these." (Note: This is also the case with boost 1.59.)

https://github.com/scantailor/scantailor/pull/194
https://github.com/scantailor/scantailor/commit/2be2b472f0cc35ef82d9e92b2813534aa780b050

Closes Homebrew/homebrew-x11#157.